app/requirements.txt
flask
mxnet
gluoncv
Pillow
numpy